scripts/requirements.txt
# Core Framework
crewai>=0.80.0
crewai-tools>=0.14.0

# LLM Providers
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.0.0

# FastAPI (for Node.js ↔ Python bridge)
fastapi>=0.115.0
uvicorn>=0.34.0

# HTTP client (for calling Node.js tool endpoints)
httpx>=0.27.0

# Settings
pydantic>=2.0.0
python-dotenv>=1.0.0
